requirements/testing.txt
-r base.txt
flake8==3.7.7 # This covers pycodestyle
pylint==2.3.1
pylint-django==2.0.9
pytest==4.6.3
pytest-django==3.5.0
pytest-cov==2.7.1
Faker==1.0.7
requests-mock==1.6.0
